Split Record into a file
