+++ ruststep/src/parser/exchange/mod.rs
@@ -4,12 +4,14 @@ mod anchor;
 mod data;
 mod header;
 mod parameter;
+mod record;
 mod reference;
 
 pub use anchor::*;
 pub use data::*;
 pub use header::*;
 pub use parameter::*;
+pub use record::*;
 pub use reference::*;
 
 use crate::parser::{combinator::*, token::*};
